Start matching navi.cpp
Got TNaviEffect and member structs finally fixed up c:

+++ include/mapCode.h
@@ -27,7 +27,7 @@ struct Code {
 	};
 #pragma enumsalwaysint reset
 
-	u8 getAttribute();
+	int getAttribute();
 	char* getAttributeName();
 	u8 getSlipCode();
 	char* getSlipCodeName();
